renamed client to demo folder, created d3 demo file folder, moved files to d3 demo from react-d3 that were d3 examples
